PM / devfreq: bw_hwmon: Rename bwmon_type enum to confuse less
It's confusing to think of the monitor register interface types
as the bw monitor type, because we already have a compatible
string in DT for the bw monitor type that isn't following the
same numbering scheme. Rename this to MON1, MON2 so that we don't
confuse it with the compatible scheme of bwmon1, bwmon2, bwmon3,
etc.

Change-Id: I406fc09ab8b9c72c43135e3deac2a382bfca9f0e
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>
Signed-off-by: celtare21 <celtare21@gmail.com>