qcom-cpufreq: Register cooling device in ready callback
Thermal cooling device has to be registered when the policy for a CPU is
ready. Cpufreq will get a callback when a cpufreq policy is ready and
register CPU cooling device as a part of this callback, so that the CPU
can be mitigated immediately if needed.

Ignore cpu cooling device registration when there is platform cooling
device available.

Change-Id: I7cfb8598aa8ead4091a617da3faddf86ff0fe6a8
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>
Signed-off-by: Manaf Meethalavalappu Pallikunhi <manafm@codeaurora.org>
Signed-off-by: YousefAlgadri <yusufgadrie@gmail.com>
Signed-off-by: dopaemon <polarisdp@gmail.com>