defconfig: implement RCU boost and NO-CB CPUs mechanisms
Latest changes in RCU should now improve the conditions for enabling
RCU boosting without much power costs.

Priority: 6.

This level of PRIO should be the proper spot on latency sensitive
workloads like a mobile platform with improved RT consumers in
userspace. A value of 2 could also be acceptable for those targets.

Delay: 250ms.

We may not need half a second to boost after the grace period expires,
reducing it to half of the default value to reduce response.

NO-CB: All [0-3 for MSM8996].

Processes RCU callbacks in different threads, in this case we allow all
the four CPUs we have available on this target to "share" their load when
high intensity RCU callbacks kick-in, a condition that could lead to RCU
starvation without it (overload of the softirq mechanism).

Change-Id: I8b74b71bc80d916f0da9fa70a5d0b7ca1b21a91c
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>

Signed-off-by: dopaemon <polarisdp@gmail.com>